Dratt ut statistikk i egen service, dratt ut mapping i egen service, lagt inn igjen getOppgave
