+++ src/main/kotlin/no/nav/klage/oppgave/domain/OppgaverSearchCriteria.kt
@@ -35,4 +35,7 @@ data class OppgaverSearchCriteria(
     enum class Projection {
         UTVIDET
     }
+
+    fun isProjectionUtvidet(): Boolean = Projection.UTVIDET == projection
+
 }
